Resolve import loop
```
    File "/root/.pyenv/versions/3.7.5/lib/python3.7/site-packages/PyInstaller/loader/pyimod03_importers.py", line 627, in exec_module
    File "site-packages/sno-0.8.1.dev0+ci.2027.git12d360fa-py3.7.egg/sno/status.py", line 8, in <module>
    File "/root/.pyenv/versions/3.7.5/lib/python3.7/site-packages/PyInstaller/loader/pyimod03_importers.py", line 627, in exec_module
    File "site-packages/sno-0.8.1.dev0+ci.2027.git12d360fa-py3.7.egg/sno/merge.py", line 6, in <module>
    File "/root/.pyenv/versions/3.7.5/lib/python3.7/site-packages/PyInstaller/loader/pyimod03_importers.py", line 627, in exec_module
    File "site-packages/sno-0.8.1.dev0+ci.2027.git12d360fa-py3.7.egg/sno/commit.py", line 24, in <module>
  ImportError: cannot import name 'get_branch_status_message' from 'sno.status' (/opt/sno/sno/status.pyc)
```

import loop was `status` --> `merge` --> `commit` --> `status`,
and only came up when you import `status` before `commit`.

Moving `merge_status_to_text` to `merge_util.py` means that `status`
doesn't have to import from `merge.py`, and the problem is resolved
